Algumas correções necessárias de acordo com o guião

+++ src/rides.h
@@ -10,9 +10,9 @@ typedef struct rides{
     Driver *driver; // driver's ID  (rework)
     User *user; // User's username (rework)
     char *city; // ride city
-    int distance;   // ride distance
-    int score_user; // user's score in this ride
-    int score_driver;   // driver's score in this ride
+    float distance;   // ride distance
+    float score_user; // user's score in this ride
+    float score_driver;   // driver's score in this ride
     float tip;  // tip from the user = not to add to trip_value
     char *comment;  // a comment
 } Ride;
